add feature, combine
- add single concep feature
- add vicinity feature
- add combine class
- reconstruct file
- add attribute assertion in Concept class

+++ src/main/java/vn/edu/hcmut/emrre/core/entity/Relation.java
@@ -128,7 +128,7 @@ public class Relation {
         return check;
     }
 
-    public static double valueOfType(Type type) {
+    public static int valueOfType(Type type) {
         switch (type) {
         case TrIP:
             return 1;
